Changes:
- move encoders2 to encoders, they are compatible
- rename json_encoder to json_encode with the old name as alias (was
  only internal, alias for e.g. esmerald)
- update release notes
- update encoders docs
