+++ lilya/encoders.py
@@ -1,3 +1,19 @@
-from lilya._internal._encoders import ENCODER_TYPES, Encoder, register_encoder
+from lilya._internal._encoders import (
+    ENCODER_TYPES,
+    Encoder,
+    EncoderProtocol,
+    WithEncodeProtocol,
+    apply_structure,
+    json_encode,
+    register_encoder,
+)
 
-__all__ = ["ENCODER_TYPES", "register_encoder", "Encoder"]
+__all__ = [
+    "ENCODER_TYPES",
+    "register_encoder",
+    "Encoder",
+    "EncoderProtocol",
+    "WithEncodeProtocol",
+    "json_encode",
+    "apply_structure",
+]
