--- lilya/encoders2.py
@@ -1,19 +1,0 @@
-from lilya._internal._encoders2 import (
-    ENCODER_TYPES,
-    Encoder,
-    EncoderProtocol,
-    WithEncodeProtocol,
-    apply_structure,
-    json_encode,
-    register_encoder,
-)
-
-__all__ = [
-    "ENCODER_TYPES",
-    "register_encoder",
-    "Encoder",
-    "EncoderProtocol",
-    "WithEncodeProtocol",
-    "json_encode",
-    "apply_structure",
-]
