+++ tests/conftest.py
@@ -1,14 +1,12 @@
 from __future__ import annotations
 
 import functools
-import sys
 from typing import Any
 
 import msgspec
 import pytest
 from attrs import asdict, has
 from msgspec import Struct
-from pydantic import BaseModel
 
 from lilya.encoders import Encoder, register_encoder
 from lilya.testclient import TestClient
@@ -21,21 +19,6 @@ def test_client_factory(anyio_backend_name, anyio_backend_options):
         backend=anyio_backend_name,
         backend_options=anyio_backend_options,
     )
-
-
-if sys.version_info <= (3, 9):  # pragma: no cover
-    dont_run = True
-else:  # pragma: no cover
-    dont_run = False
-
-    class PydanticEncoder(Encoder):
-        def is_type(self, value: Any) -> bool:
-            return isinstance(value, BaseModel)
-
-        def serialize(self, obj: BaseModel) -> dict[str, Any]:
-            return obj.model_dump()
-
-    register_encoder(PydanticEncoder())
 
 
 class MsgSpecEncoder(Encoder):
